Added utils.py
utils.py will contain scripts for invoking mid3v2 and vorbiscomment
on the correct files. Unlike the perl implementation, the plan is to
only run these commands on .oggs and mp3s instead of every file in the
directory. This way, there wont be a bunch of empty songs parsed by the
MID and Vorbis parsers.